requirements.txt
Django==1.11.29
django-extensions==1.9.8
#git+https://github.com/xzzy/django-proxy.git#egg=proxy
git+https://github.com/dbca-wa/wagov_utils.git#egg=wagov_utils
git+https://github.com/dbca-wa/ledger#egg=ledger
#git+https://github.com/dbca-wa/ledger@v2022.08.19.01#egg=ledger
#git+https://github.com/dbca-wa/ledger@4818666a1bda5fac274e9d0b48008797f04f6545
git+https://github.com/dbca-wa/dpaw-utils.git@0.3a16#egg=dpaw-utils
git+https://github.com/django-oscar/django-oscar.git@8a3288da439cc2a878f44ae5c5101043e658d2a2#egg=django-oscar
social-auth-app-django==2.1.0
git+https://github.com/scottp-dpaw/social-core.git@email_fix#egg=social-auth-core
coverage==4.3.1
coveralls==1.1
reportlab==3.5.55
django_bootstrap3==12.1.0
django-datatables-view==1.13.0
django-reversion==3.0.0
django-reversion-compare==0.8.6
git+https://github.com/xzzy/django-preserialize.git#egg=django-preserialize
django-countries==6.1.3
django-cron==0.5.0
django-dynamic-fixture==1.9.1
openpyxl==3.0.9
datapackage==0.8.1
py4j==0.10.2.1
djangorestframework==3.5.2
djangorestframework-gis==0.11
djangorestframework-csv==2.0.0
pycountry==17.1.8
six>=1.10.0
django-ical==1.4
django-taggit==0.21.3
geojson==1.3.3
unicodecsv==0.14.1
PyPDF2==1.26.0
django-extra-fields==0.9
django-crispy-forms==1.7.0
django-reset-migrations==0.3.1
django-ckeditor==5.4.0
djangorestframework-datatables==0.4.0
django-confy==1.0.4
xlwt==1.3.0
xlsxwriter==1.1.2
#gunicorn==19.9.0
gunicorn==23.0.0
dj-static==0.0.6
mixer==6.1.3
django-dirtyfields==1.3.1 
dj-database-url==0.5.0
python-docx==0.8.10
docxtpl==0.10.0
phonenumbers==8.12.15
django-treebeard==4.3.1
git+https://github.com/digi604/django-smart-selects.git@js-unlinting-fixes
django-multiselectfield==0.1.12
psycopg2==2.8.6
deepdiff==5.7.0
whitenoise==5.3.0
git+https://github.com/dbca-wa/django-media-serv.git#egg=django_media_serv
drf-writable-nested==0.7.0

# next packages required for geopandas
# rtree requires 'sudo apt install libspatialindex-dev'
click-plugins==1.1.1
cligj==0.7.2
fiona==1.8.21
geopandas==0.10.2
munch==2.5.0
#numpy==1.21.5
numpy==1.18.5
pandas==1.3.5
pyproj==3.2.1
python-dateutil==2.8.2
shapely==1.8.0
geopandas==0.10.2
rtree==0.9.7

git+https://github.com/dbca-wa/appmonitor_client.git#egg=appmonitor_client
python-decouple==3.8
